Fix GitHub timezone issue
According to the following article from GitHub blog:
Timezone-aware contribution graphs
https://github.com/blog/1793-timezone-aware-contribution-graphs
after Monday 10 March 2014, when counting commits, they use the timezone information present in the timestamps for those commits.
But before Monday 10 March 2014, they use the server timezone which is UTC-8 (got that info from google). In this commit the date before Monday 10 March 2014 is adjusted to fit the situation.